native font style

+++ api/base/src/lib.rs
@@ -53,9 +53,9 @@ impl Bounds {
 
 #[derive(Debug, Clone)]
 pub struct Color {
-	pub r: f32,
-	pub g: f32,
-	pub b: f32,
+	pub r: u8,
+	pub g: u8,
+	pub b: u8,
 	pub a: f32,
 }
